Split benchmarks for convolution operators
